+++ packages/shared/database/migrations.ts
@@ -10,6 +10,7 @@ import * as _006 from './migrations/006-survey-reddit-dataset.js'
 import * as _007 from './migrations/007-nullable-participant-id.js'
 import * as _008 from './migrations/008-respone-new-fields.js'
 import * as _009 from './migrations/009-responses-created-at.js'
+import * as _010 from './migrations/010-survey-updates.js'
 
 class Migrations implements MigrationProvider {
   async getMigrations(): Promise<Record<string, Migration>> {
@@ -23,6 +24,7 @@ class Migrations implements MigrationProvider {
       '007-nullable-participant-id': _007,
       '008-respone-new-fields': _008,
       '009-responses-created-at': _009,
+      '010-survey-updates': _010,
     }
   }
 }