+++ packages/worker/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @credee/worker
 
+## 0.2.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @credee/shared@0.4.0
+
 ## 0.2.2
 
 ### Patch Changes